Improve IDN event tests and fix warnings
