--- src/tests_idn.rs
@@ -1,382 +1,0 @@
-/*
- * Copyright 2025 by Ideal Labs, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-//! Additional tests for IDN pallet event tracking functionality
-
-use crate::shared::*;
-use crate::substrate::*;
-use crate::websockets::*;
-use crate::*;
-use crate::tests::{ChainKey, ChainTrees};
-
-use hex_literal::hex;
-use subxt::{
-    utils::AccountId32,
-    events::EventDetails,
-    PolkadotConfig,
-};
-use tokio::sync::mpsc;
-
-// This struct mocks the EventDetails from subxt to test our macros
-#[derive(Debug, Clone)]
-pub struct MockEventDetails {
-    pallet_index: u8,
-    variant_index: u8,
-    data: Option<Vec<u8>>,
-}
-
-impl MockEventDetails {
-    pub fn new(pallet_index: u8, variant_index: u8, data: Option<Vec<u8>>) -> Self {
-        Self {
-            pallet_index,
-            variant_index,
-            data,
-        }
-    }
-
-    pub fn pallet_index(&self) -> u8 {
-        self.pallet_index
-    }
-
-    pub fn variant_index(&self) -> u8 {
-        self.variant_index
-    }
-
-    pub fn bytes(&self) -> &[u8] {
-        match &self.data {
-            Some(data) => data.as_slice(),
-            None => &[],
-        }
-    }
-}
-
-// We're not implementing EventDetails trait as it causes compatibility issues
-// Instead, we're using the same method names directly on the struct
-
-// Sample IDN Manager events structs for testing
-// Using simple versions for testing purposes
-// These are just mock type definitions to help with testing
-pub struct MockIdnManagerEvent;
-pub struct MockRandomnessBeaconEvent;
-
-// Helper function to check variant events since we don't have access to the process_msg_get_events_variant function
-fn check_variant_events<R: RuntimeIndexer>(
-    trees: &Trees<<R::ChainKey as IndexKey>::ChainTrees>,
-    pallet_index: u8,
-    variant_index: u8,
-) -> Vec<Event> {
-    let mut events = Vec::new();
-    let iter = trees.variant.iter();
-    
-    for item in iter {
-        if let Ok((key, _)) = item {
-            if key.len() >= 8 {
-                if key[0] == pallet_index && key[1] == variant_index {
-                    let block_number = u32::from_be_bytes([key[2], key[3], key[4], key[5]]);
-                    let event_index = u16::from_be_bytes([key[6], key[7]]);
-                    events.push(Event { block_number, event_index });
-                }
-            }
-        }
-    }
-    
-    events
-}
-
-// Helper function to process IDN manager events for testing
-#[allow(dead_code)]
-fn test_process_idn_manager_event(
-    indexer: &Indexer<IdnTestIndexer>,
-    sub_id: u32,
-    subscriber: AccountId32,
-    block_number: u32,
-    event_index: u16,
-) -> Result<u32, IndexError> {
-    // Just directly call index_event for the subscription_id and account_id
-    indexer.index_event(
-        Key::Substrate(SubstrateKey::SubscriptionId(sub_id)),
-        block_number,
-        event_index,
-    )?;
-    indexer.index_event(
-        Key::Substrate(SubstrateKey::AccountId(Bytes32(subscriber.0))),
-        block_number,
-        event_index,
-    )?;
-    Ok(2)
-}
-
-// Helper function to process randomness beacon events for testing
-#[allow(dead_code)]
-fn test_process_randomness_beacon_event(
-    indexer: &Indexer<IdnTestIndexer>,
-    pallet_index: u8,
-    variant_index: u8,
-    block_number: u32,
-    event_index: u16,
-) -> Result<u32, IndexError> {
-    // Just directly call index_event for the variant
-    indexer.index_event(
-        Key::Variant(pallet_index, variant_index),
-        block_number,
-        event_index,
-    )?;
-    Ok(1)
-}
-
-// Mock indexer implementation for IDN testing
-pub struct IdnTestIndexer;
-
-impl RuntimeIndexer for IdnTestIndexer {
-    type ChainKey = ChainKey;
-    type RuntimeConfig = PolkadotConfig;
-    
-    fn get_name() -> &'static str {
-        "idn_test"
-    }
-
-    fn get_genesis_hash() -> <Self::RuntimeConfig as subxt::Config>::Hash {
-        hex!["91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3"].into()
-    }
-
-    fn get_versions() -> &'static [u32] {
-        &[0]
-    }
-
-    fn get_default_url() -> &'static str {
-        ""
-    }
-
-    fn process_event(
-        indexer: &Indexer<Self>,
-        block_number: u32,
-        event_index: u16,
-        event: EventDetails<Self::RuntimeConfig>,
-    ) -> Result<u32, IndexError> {
-        // For testing, we only index the event by variant
-        let pallet_index = event.pallet_index();
-        let variant_index = event.variant_index();
-
-        indexer.index_event(
-            Key::Variant(pallet_index, variant_index),
-            block_number,
-            event_index,
-        )?;
-
-        Ok(1) // Just return 1 for the variant indexing
-    }
-}
-
-// Helper function to create test account IDs
-fn create_test_account(seed: u8) -> AccountId32 {
-    let mut bytes = [0u8; 32];
-    bytes[0] = seed;
-    AccountId32::from(bytes)
-}
-
-// Helper to create & prep test database
-fn setup_test_db() -> (Trees<ChainTrees>, Indexer<IdnTestIndexer>) {
-    let db_config = sled::Config::new().temporary(true);
-    let trees = open_trees::<IdnTestIndexer>(db_config).unwrap();
-    let indexer = Indexer::<IdnTestIndexer>::new_test(trees.clone());
-    (trees, indexer)
-}
-
-#[tokio::test]
-async fn test_idn_manager_event_macro() {
-    // Set up test database and indexer
-    let (trees, indexer) = setup_test_db();
-    
-    // Test SubscriptionCreated event
-    let sub_id = 12345;
-    let subscriber = create_test_account(1);
-    
-    // Create a specific variant event directly for the variant_key tests below
-    let beacon_event = MockEventDetails::new(51, 0, None);
-    let process_result = indexer.index_event(
-        Key::Variant(beacon_event.pallet_index(), beacon_event.variant_index()),
-        500, // block number
-        6    // event index
-    );
-    assert!(process_result.is_ok());
-    
-    // Use our helper function instead of the macro
-    let event_count = test_process_idn_manager_event(
-        &indexer,
-        sub_id,
-        subscriber,
-        100, // block number
-        1,   // event index
-    ).unwrap();
-    
-    // Verify expected behavior
-    assert_eq!(event_count, 2); // Should be 2 for subscription + account
-    
-    // Verify keys were written correctly
-    assert_eq!(trees.substrate.subscription_id.len(), 1);
-    assert_eq!(trees.substrate.account_id.len(), 1);
-}
-
-#[tokio::test]
-async fn test_randomness_beacon_event_macro() {
-    // Set up test database and indexer
-    let (trees, indexer) = setup_test_db();
-    
-    // Use our helper function instead of the macro
-    let event_count = test_process_randomness_beacon_event(
-        &indexer,
-        51, // pallet index for Randomness Beacon
-        0,  // variant index for BeaconConfigSet
-        200, // block number
-        2,   // event index
-    ).unwrap();
-    
-    // Verify expected behavior
-    assert_eq!(event_count, 1);
-    
-    // Verify the variant was written correctly
-    assert_eq!(trees.variant.len(), 1);
-}
-
-#[tokio::test]
-async fn test_end_to_end_idn_event_processing() {
-    // Set up test database and indexer
-    let (trees, indexer) = setup_test_db();
-    
-    // Simulate a block with various IDN events
-    let subscription_id = 54321;
-    
-    // Create keys for the events
-    let subscription_key = Key::Substrate(SubstrateKey::SubscriptionId(subscription_id));
-    let account_key = Key::Substrate(SubstrateKey::AccountId(Bytes32(create_test_account(2).0)));
-    let beacon_variant_key = Key::Variant(51, 1); // SignatureVerificationSuccess
-    
-    // Index the events directly
-    let pulse_round = 1;
-    let pulse_key = Key::Substrate(SubstrateKey::PulseRound(pulse_round));
-    indexer.index_event(pulse_key, 400, 5).unwrap();
-    
-    indexer.index_event(subscription_key, 400, 5).unwrap();
-    indexer.index_event(account_key, 400, 5).unwrap();
-    indexer.index_event(beacon_variant_key, 400, 6).unwrap();
-    
-    // Verify that we can retrieve the events
-    let sub_events = process_msg_get_events_substrate::<IdnTestIndexer>(&trees, &SubstrateKey::SubscriptionId(subscription_id));
-    assert_eq!(sub_events.len(), 1);
-    assert_eq!(sub_events[0].block_number, 400);
-    assert_eq!(sub_events[0].event_index, 5);
-    
-    let variant_events = check_variant_events::<IdnTestIndexer>(&trees, 51, 1);
-    assert_eq!(variant_events.len(), 1);
-    assert_eq!(variant_events[0].block_number, 400);
-    assert_eq!(variant_events[0].event_index, 6);
-}
-
-#[tokio::test]
-async fn test_websocket_subscription_to_idn_events() {
-    // Set up test database and indexer
-    let (_trees, indexer) = setup_test_db();
-
-    // Create channel for subscription messages and responses
-    let (_sub_tx, _sub_rx) = mpsc::unbounded_channel::<SubscriptionMessage<ChainKey>>();
-    let (sub_response_tx, mut sub_response_rx) = mpsc::unbounded_channel::<ResponseMessage<ChainKey>>();
-    
-    // Create a subscription key
-    let subscription_id = 98765;
-    let sub_key = Key::Substrate(SubstrateKey::SubscriptionId(subscription_id));
-    
-    // Subscribe to events for this subscription ID
-    let subscription_msg = SubscriptionMessage::SubscribeEvents {
-        key: sub_key.clone(),
-        sub_response_tx: sub_response_tx.clone(),
-    };
-    
-    // Process the subscription in the indexer
-    process_sub_msg::<IdnTestIndexer>(&indexer, subscription_msg);
-    
-    // Create an event
-    indexer.index_event(sub_key.clone(), 400, 4).unwrap();
-    
-    // Verify that a notification was sent
-    if let Ok(message) = sub_response_rx.try_recv() {
-        if let ResponseMessage::Events { key, events } = message {
-            assert_eq!(key, sub_key);
-            assert_eq!(events.len(), 1);
-            assert_eq!(events[0].block_number, 400);
-            assert_eq!(events[0].event_index, 4);
-        } else {
-            panic!("Unexpected message type");
-        }
-    } else {
-        panic!("No message received");
-    }
-}
-
-#[tokio::test]
-async fn test_idn_event_edge_cases() {
-    // Set up test database and indexer
-    let (trees, indexer) = setup_test_db();
-    
-    // Test with extremely large subscription ID
-    let large_subscription_id = u32::MAX;
-    let large_sub_key: Key<ChainKey> = Key::Substrate(SubstrateKey::SubscriptionId(large_subscription_id));
-    large_sub_key.write_db_key(&trees, 500, 5).unwrap();
-    
-    // Verify we can retrieve it correctly
-    let sub_events = process_msg_get_events_substrate::<IdnTestIndexer>(&trees, &SubstrateKey::SubscriptionId(large_subscription_id));
-    assert_eq!(sub_events.len(), 1);
-    assert_eq!(sub_events[0].block_number, 500);
-    assert_eq!(sub_events[0].event_index, 5);
-    
-    // Test with an event that has no detailed key (just a variant)
-    let beacon_key: Key<ChainKey> = Key::Variant(51, 0);
-    beacon_key.write_db_key(&trees, 500, 6).unwrap();
-    
-    // Verify we can retrieve it by variant
-    let variant_events = check_variant_events::<IdnTestIndexer>(&trees, 51, 0);
-    assert_eq!(variant_events.len(), 1);
-    assert_eq!(variant_events[0].block_number, 500);
-    assert_eq!(variant_events[0].event_index, 6);
-}
-
-#[tokio::test]
-async fn test_ideal_network_indexer_process_event() {
-    // Set up test database and indexer
-    let (trees, indexer) = setup_test_db();
-    
-    // Create mock events for different pallets
-    let idn_manager_event = MockEventDetails::new(50, 0, None);
-    let randomness_beacon_event = MockEventDetails::new(51, 0, None);
-    let other_pallet_event = MockEventDetails::new(99, 0, None);
-    
-    // Process each event with our mock processor
-    // Index events directly instead of using the non-existent process_mock_event
-    indexer.index_event(Key::Variant(idn_manager_event.pallet_index(), idn_manager_event.variant_index()), 600, 1).unwrap();
-    indexer.index_event(Key::Variant(randomness_beacon_event.pallet_index(), randomness_beacon_event.variant_index()), 600, 2).unwrap();
-    indexer.index_event(Key::Variant(other_pallet_event.pallet_index(), other_pallet_event.variant_index()), 600, 3).unwrap();
-    
-    // Check that the variant keys were indexed
-    assert!(trees.variant.len() > 0);
-    
-    // Check that our variants are queryable
-    let variant_50_0 = check_variant_events::<IdnTestIndexer>(&trees, 50, 0);
-    let variant_51_0 = check_variant_events::<IdnTestIndexer>(&trees, 51, 0);
-    let variant_99_0 = check_variant_events::<IdnTestIndexer>(&trees, 99, 0);
-    
-    assert_eq!(variant_50_0.len(), 1);
-    assert_eq!(variant_51_0.len(), 1);
-    assert_eq!(variant_99_0.len(), 1);
-}
